requirements.txt
asgi-ipc==1.3.1
asgi-redis==1.2.1
asgiref==1.1.1
asn1crypto==0.24.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
autopep8==1.3.5
captcha==0.2.4
certifi==2018.4.16
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
cloudinary==1.12.0
constantly==15.1.0
cryptography==2.3.1
daphne==1.4.2
defusedxml==0.5.0
Django==1.11
django-countries==5.3.1
django-environ==0.4.5
django-filter==2.0.0
django-filters==0.2.1
django-material==1.2.6
django-recaptcha==1.4.0
django-render-block==0.5
django-richtextfield==1.2.1
django-templated-email==2.2.0
django-timezone-field==2.1
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
mock==2.0.0
msgpack-python==0.5.6
oauthlib==2.1.0
paypalrestsdk==1.13.1
pbr==4.0.4
Pillow==5.2.0
posix-ipc==1.0.4
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pycparser==2.18
PyHamcrest==1.9.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
stevedore==1.28.0
Twisted==18.7.0
txaio==2.10.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
zope.interface==4.5.0
